feat: реализация сервисов order, payment и delivery для интернет-магазина
- Добавлен микросервис order с логикой создания и управления заказами
- Реализован сервис payment с расчётом стоимости товаров, доставки и налога (НДС 10%)
- Добавлен сервис delivery с расчётом стоимости доставки и управлением статусами
- Настроено взаимодействие сервисов через Feign-клиенты
- Добавлена интеграция с warehouse и shopping-store
- Подготовлена базовая поддержка Gateway и Eureka (order/payment/delivery)

+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/dto/cart/ChangeProductQuantityRequest.java
@@ -16,5 +16,5 @@ public class ChangeProductQuantityRequest {
     private UUID productId;
     @NotNull
     @Min(1)
-    private Integer newQuantity;
+    private Long newQuantity;
 }
